% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_symbols.R
\name{get_symbols}
\alias{get_symbols}
\title{Get tse data}
\usage{
get_symbols(symbol = NULL, env = NULL, auto_assign = FALSE,
  verbose = FALSE, return_class = c("tibble", "xts"),
  adj_prices = FALSE)
}
\arguments{
\item{adj_prices}{}
}
\description{
Get tse data
}
