% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TseClient.R
\docType{package}
\name{TseClient-package}
\alias{TseClient}
\alias{TseClient-package}
\title{TseClient: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (YOUR-ORCID-ID)

}
